Remove old ASM implementations
