+++ gradle.properties
@@ -70,7 +70,7 @@ mixinConfigRefmap=
 generateMixinConfig = false
 # Specify the core mod entry class if you use a core mod. This class must implement IFMLLoadingPlugin!
 # Example value: coreModClass = asm.FMLPlugin + modGroup = com.myname.mymodid -> com.myname.mymodid.asm.FMLPlugin
-coreModClass = asm.GregTechLoadingPlugin
+coreModClass=mixins.GregTechMixinLoadingPlugin
 # If your project is only a consolidation of mixins or a core mod and does NOT contain a 'normal' mod (meaning that
 # there is no class annotated with @Mod) you want this to be true. When in doubt: leave it on false!
 containsMixinsAndOrCoreModOnly=false
