--- src/main/java/gregtech/asm/hooks/RenderChunkHooks.java
@@ -1,22 +1,0 @@
-package gregtech.asm.hooks;
-
-import gregtech.api.metatileentity.MetaTileEntityHolder;
-
-import net.minecraft.client.renderer.tileentity.TileEntityRendererDispatcher;
-import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
-import net.minecraft.tileentity.TileEntity;
-
-import org.jetbrains.annotations.Nullable;
-
-@SuppressWarnings("unused")
-public class RenderChunkHooks {
-
-    public static <T extends TileEntity> TileEntitySpecialRenderer<T> getRenderer(TileEntityRendererDispatcher renderer,
-                                                                                  @Nullable TileEntity tileEntityIn) {
-        // TODO
-        if (tileEntityIn instanceof MetaTileEntityHolder && !((MetaTileEntityHolder) tileEntityIn).hasTESR()) {
-            return null;
-        }
-        return renderer.getRenderer(tileEntityIn);
-    }
-}
