--- src/main/java/gregtech/asm/package-info.java
@@ -1,6 +1,0 @@
-/**
- * This package contains the GregTech: CE Unofficial CoreMod.
- *
- * If you are importing anything from this package, you are probably looking in the wrong place.
- */
-package gregtech.asm;
