+++ src/main/java/gregtech/client/utils/CTMHooks.java
@@ -1,7 +1,7 @@
-package gregtech.asm.hooks;
+package gregtech.client.utils;
 
+import gregtech.api.util.GTLog;
 import gregtech.api.util.Mods;
-import gregtech.client.utils.BloomEffectUtil;
 
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.block.model.BakedQuad;
@@ -10,12 +10,26 @@ import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumFacing;
 import net.minecraftforge.client.ForgeHooksClient;
 
+import team.chisel.ctm.api.model.IModelCTM;
+import team.chisel.ctm.client.model.ModelCTM;
+
+import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-@SuppressWarnings("unused")
 public class CTMHooks {
+
+    private static Field layers;
+
+    static {
+        try {
+            layers = ModelCTM.class.getDeclaredField("layers");
+            layers.setAccessible(true);
+        } catch (NoSuchFieldException e) {
+            GTLog.logger.error("CTMModHooks no such field");
+        }
+    }
 
     public static ThreadLocal<Boolean> ENABLE = new ThreadLocal<>();
 
@@ -48,5 +62,18 @@ public class CTMHooks {
             }
         }
         return ret;
+    }
+
+    public static boolean canRenderInLayer(IModelCTM model, IBlockState state, BlockRenderLayer layer) {
+        boolean canRenderInLayer = model.canRenderInLayer(state, layer);
+        if (model instanceof ModelCTM && layers != null) {
+            try {
+                return CTMHooks.checkLayerWithOptiFine(canRenderInLayer, layers.getByte(model), layer);
+            } catch (Exception ignored) {
+                layers = null;
+                GTLog.logger.error("CTMHooks Field error");
+            }
+        }
+        return canRenderInLayer;
     }
 }
